Data type and saturation fix for MedullaEncoder.

+++ atrias-robot/MedullaEncoder.m
@@ -26,11 +26,11 @@ classdef MedullaEncoder < Encoder
 	methods (Access = protected)
 		function [pos, vel, isValid] = stepImpl(this, ticks, counter, timestamp, calibTicks, calibVal, calibTrig, unitsPerTick, unwrapMod, minPos, maxPos, minVel, maxVel)
 			% Compute the delta time value for this iteration
-			dt = this.sample_time * mod(counter - this.prevCntr, this.counterMod) + (timestamp - this.prevTs) / this.medullaFreq;
+			dt = this.sample_time * mod(double(counter) - this.prevCntr, this.counterMod) + (double(timestamp) - this.prevTs) / this.medullaFreq;
 
 			% Update the stored values used in the above calculation
-			this.prevCntr = counter;
-			this.prevTs   = timestamp;
+			this.prevCntr = double(counter);
+			this.prevTs   = double(timestamp);
 
 			% Forward the remaining work to the Encoder class
 			[pos, vel, isValid] = stepImpl@Encoder(this, ticks, dt, calibTicks, calibVal, calibTrig, unitsPerTick, unwrapMod, minPos, maxPos, minVel, maxVel);
